Pruning/pyproject.toml
[project]
name = "pruning-bootcamp"
version = "0.1.0"
description = "21-Day Neural Network Pruning Bootcamp - From Unstructured to Structured Pruning"
readme = "README.md"
requires-python = "==3.11.*"

dependencies = [
  "numpy~=2.0",
  "pandas~=2.2",
  "matplotlib~=3.8",
  "seaborn~=0.13",
  "tqdm>=4.66,<5",
  "pyyaml>=6,<7",
  "tensorboard>=2.13,<3",
  "onnx>=1.14,<1.18",
  "torch-pruning>=1.3,<2",
]

[project.optional-dependencies]
cpu = [
  "torch~=2.7",
  "torchvision~=0.20",
  "onnxruntime>=1.20,<1.25",
  "thop>=0.1,<1",
]

cuda126 = [
  "torch~=2.7; extra == 'cuda126'",
  "torchvision~=0.20; extra == 'cuda126'",
  "onnxruntime-gpu>=1.20,<1.25",
  "thop>=0.1,<1",
]

cuda128 = [
  "torch~=2.7; extra == 'cuda128'",
  "torchvision~=0.20; extra == 'cuda128'",
  "onnxruntime-gpu>=1.20,<1.25",
  "thop>=0.1,<1",
]

cuda129 = [
  "torch~=2.7; extra == 'cuda129'",
  "torchvision~=0.20; extra == 'cuda129'",
  "onnxruntime-gpu>=1.20,<1.25",
  "thop>=0.1,<1",
]

tvm = [
  "apache-tvm>=0.16,<1.dev0; python_version >= '3.12'"
]

dev = [
  "pytest~=8.3",
  "pytest-cov~=5.0",
  "ruff>=0.5,<1",
  "mypy>=1.10,<2",
]

[tool.pytest.ini_options]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true